json for genesis, assets, new component for syntax highlighting
