+++ apps/explorer/src/app/routes/genesis/index.tsx
@@ -1,15 +1,17 @@
-import { DATA_SOURCES } from "../../config";
-import useFetch from "../../hooks/use-fetch";
-import { TendermintGenesisResponse } from "./tendermint-genesis-response";
+import { SyntaxHighlighter } from '../../components/syntax-highlighter';
+import { DATA_SOURCES } from '../../config';
+import useFetch from '../../hooks/use-fetch';
+import { TendermintGenesisResponse } from './tendermint-genesis-response';
 
 const Genesis = () => {
   const { data: genesis } = useFetch<TendermintGenesisResponse>(
     `${DATA_SOURCES.tendermintUrl}/genesis`
   );
+  if (!genesis?.result.genesis) return null;
   return (
     <section>
       <h1>Genesis</h1>
-      <pre>{JSON.stringify(genesis, null, "  ")}</pre>
+      <SyntaxHighlighter data={genesis?.result.genesis} />
     </section>
   );
 };
