+++ apps/explorer/src/app/routes/markets/index.tsx
@@ -1,8 +1,8 @@
 import { gql, useQuery } from '@apollo/client';
 import { MarketsQuery } from './__generated__/MarketsQuery';
 
-import SyntaxHighlighter from 'react-syntax-highlighter';
 import React from 'react';
+import { SyntaxHighlighter } from '../../components/syntax-highlighter';
 
 const MARKETS_QUERY = gql`
   query MarketsQuery {
@@ -144,28 +144,6 @@ const MARKETS_QUERY = gql`
   }
 `;
 
-const vegaJsonTheme = {
-  hljs: {
-    fontSize: '1rem',
-    fontFamily: "Consolas, Monaco, 'Andale Mono', 'Ubuntu Mono', monospace",
-    display: 'block',
-    overflowX: 'auto',
-    padding: '1em',
-    color: '#26ff8a',
-    background: '#2C2C2C',
-    border: '1px solid #696969',
-  },
-  'hljs-literal': {
-    color: '#ff2d5e',
-  },
-  'hljs-number': {
-    color: '#ff7a1a',
-  },
-  'hljs-string': {
-    color: '#48aff0',
-  },
-};
-
 const Markets = () => {
   const { data } = useQuery<MarketsQuery>(MARKETS_QUERY);
 
@@ -176,9 +154,7 @@ const Markets = () => {
       {data.markets.map((m) => (
         <React.Fragment key={m.id}>
           <h2>{m.name}</h2>
-          <SyntaxHighlighter language="json" style={vegaJsonTheme}>
-            {JSON.stringify(m, null, '  ')}
-          </SyntaxHighlighter>
+          <SyntaxHighlighter data={m} />
         </React.Fragment>
       ))}
     </section>
